# voice-customer-service-lab/packages/contracts/openapi/voice-api.v1.json
{
  "components": {
    "schemas": {
...
      "def-28": {
        "anyOf": [
          {
            "enum": [
              "rtc_join"
            ],
            "type": "string"
          },
          {
            "enum": [
              "turn_first_output"
            ],
            "type": "string"
          },
          {
            "enum": [
              "barge_in_stop"
            ],
            "type": "string"
          }
        ],
        "title": "RealtimeSliName"
      },
      "def-29": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "required": [
          "accepted",
          "replayed"
        ],
        "title": "RealtimeSliObservationAck",
        "type": "object"
      },
...
      "def-30": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "maximum": 60000,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "observation_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^obs_[A-Za-z0-9]+$",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "enum": [
                  "success"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "failure"
                ],
                "type": "string"
              }
            ]
          },
          "sli": {
            "$ref": "#/components/schemas/def-28"
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "mock"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "rtc"
                ],
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "observation_id",
          "sli",
          "source",
          "outcome",
          "duration_ms"
        ],
        "title": "RealtimeSliObservationRequest",
        "type": "object"
      },
...
      "def-7": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "pattern": "^[A-Z][A-Z0-9_]*$",
                "type": "string"
              },
              "correlation_id": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "retryable",
              "correlation_id"
            ],
            "title": "ApiErrorDetail",
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Voice Customer Service API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/sessions/{session_id}/realtime-observations": {
      "post": {
        "operationId": "recordRealtimeSliObservation",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-30"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-29"
                }
              }
            },
            "description": "Default Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-29"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "observability"
        ]
      }
    },
...
  }
}